+++ packages/template-retail-react-app/app/components/search/partials/search-suggestions.jsx
@@ -15,7 +15,10 @@ const SearchSuggestions = ({recentSearches, searchSuggestions, closeAndNavigate}
     const hasCategories = searchSuggestions?.categorySuggestions?.length
     const hasProducts = searchSuggestions?.productSuggestions?.length
     const hasBrands = searchSuggestions?.brandSuggestions?.length
-    const hasSuggestions = hasCategories || hasProducts || hasBrands
+    const hasPopularSearches = searchSuggestions?.popularSearchSuggestions?.length
+    const hasRecentSearches = searchSuggestions?.recentSearchSuggestions?.length
+    const hasSuggestions =
+        hasCategories || hasProducts || hasBrands || hasPopularSearches || hasRecentSearches
 
     return (
         <Stack {...styles.container}>